Функция для расчета продолжительности события и форматирования вынесена в отдельный файл
